pyproject.toml
[project]
name = "gh-issues-dc-forums-bridge"
version = "0.0.1"
description = ""
authors = [
    {name = "Nucceteere",email = "ruzgar@nucceteere.xyz"}
]
license = {text = "GPL-3.0-only"}
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "discord-py (>=2.5.2,<3.0.0)",
    "pygithub (>=2.6.1,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "psycopg[c,pool] (>=3.2.9,<4.0.0)",
    "yoyo-migrations (>=9.0.0,<10.0.0)",
]

[tool.poetry]
package-mode = false

[tool.poetry.group.dev.dependencies]
pre-commit = "^4.2.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
